feat: trigger auto mint doi on (publish and claim verification)
