+++ desci-server/src/routes/v1/admin/index.ts
@@ -3,7 +3,7 @@ import { Router } from 'express';
 import { createCsv, getAnalytics } from '../../../controllers/admin/analytics.js';
 import { listAttestations } from '../../../controllers/admin/communities/index.js';
 import { debugAllNodesHandler, debugNodeHandler } from '../../../controllers/admin/debug.js';
-import { listDoiRecords } from '../../../controllers/doi/admin.js';
+import { listDoiRecords, mintDoi } from '../../../controllers/admin/doi/index.js';
 import { ensureAdmin } from '../../../middleware/ensureAdmin.js';
 import { ensureUser } from '../../../middleware/permissions.js';
 import { asyncHandler } from '../../../utils/asyncHandler.js';
@@ -15,7 +15,9 @@ const router = Router();
 
 router.get('/analytics', [ensureUser, ensureAdmin], getAnalytics);
 router.get('/analytics/csv', [ensureUser, ensureAdmin], createCsv);
+
 router.get('/doi/list', [ensureUser, ensureAdmin], listDoiRecords);
+router.post('/mint/:uuid', [ensureUser, ensureAdmin], asyncHandler(mintDoi));
 
 router.get('/debug', [ensureUser, ensureAdmin], asyncHandler(debugAllNodesHandler));
 router.get('/debug/:uuid', [ensureUser, ensureAdmin], asyncHandler(debugNodeHandler));
