+++ desci-server/src/routes/v1/doi.ts
@@ -1,7 +1,6 @@
 import { Router } from 'express';
 
 import { checkMintability, retrieveDoi } from '../../controllers/doi/check.js';
-import { mintDoi } from '../../controllers/doi/mint.js';
 import { retrieveDoiSchema } from '../../controllers/doi/schema.js';
 import { ensureNodeAccess } from '../../middleware/authorisation.js';
 import { ensureUser } from '../../middleware/permissions.js';
@@ -11,7 +10,6 @@ import { asyncHandler } from '../../utils/asyncHandler.js';
 const router = Router();
 
 router.get('/check/:uuid', [ensureUser, ensureNodeAccess], asyncHandler(checkMintability));
-router.post('/mint/:uuid', [ensureUser, ensureNodeAccess], asyncHandler(mintDoi));
 router.get('/', [validate(retrieveDoiSchema)], asyncHandler(retrieveDoi));
 
 export default router;
