+++ desci-server/src/services/Doi.ts
@@ -1,5 +1,6 @@
 import { PdfComponent, ResearchObjectComponentType, ResearchObjectV1 } from '@desci-labs/desci-models';
 import { DoiStatus, DoiSubmissionQueue, NodeVersion, Prisma, PrismaClient } from '@prisma/client';
+// import _ from 'lodash';
 import { v4 } from 'uuid';
 
 import {
@@ -206,6 +207,17 @@ export class DoiService {
 
     // return submission queue data
     return submission;
+  }
+
+  async autoMintTrigger(uuid: string) {
+    const sanitizedUuid = ensureUuidEndsWithDot(uuid);
+    const isPending = await this.hasPendingSubmission(sanitizedUuid);
+    if (isPending) {
+      throw new MintError('You have a pending submission');
+    } else {
+      const submission = await this.mintDoi(sanitizedUuid);
+      return submission;
+    }
   }
 
   /**
