+++ desci-server/src/services/NotificationService.ts
@@ -397,7 +397,7 @@ export const emitNotificationOnAttestationValidation = async ({
   const claim = await attestationService.findClaimById(claimId);
   const versionedAttestation = await attestationService.getAttestationVersion(claim.attestationVersionId, claimId);
   const dpid = await getDpidFromNode(node);
-
+  logger.trace({ versionedAttestation, claim, claimId }, 'INSPECT CLAIM');
   const attestationName = versionedAttestation.name;
 
   const payload: AttestationValidationPayload = {
